+++ app/views/steps/check/results/show.en.html.erb
@@ -1,4 +1,4 @@
-<% title '' %>
+<% title t('.page_title') %>
 
 <div class="govuk-width-container">
   <main class="govuk-main-wrapper" id="main-content">
